Support for Script introductions
Similar to the ones found in the base 3 almanacs, script writers may have introductions they want to give, or possibly instructions to the players for special rules (i.e. Stormcatcher being expected to favor X character). This gives the ability to add an introduction like that to your script. An example of Trouble Brewing has been provided with the introduction in that almanac.

Also moved the images to a sub directory within Images as a first move to homebrew support.